pyproject.toml
[project]
name = "rbc-pdf-to-csv"
version = "0.1.0"
description = "CLI tool to convert bank statement PDFs to CSV files"
authors = [
    { name = "Colin Bendell", email = "colin@bendell.ca" }
]
requires-python = ">=3.12"
dependencies = [
    "pymupdf>=1.26.0",
    "click>=8.1.7",
    "pandas>=2.2.0",
    "anthropic>=0.47.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
]

[project.scripts]
rbc-pdf-to-csv = "rbc2.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rbc2"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.6.1",
    "ruff>=0.14.5",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers -n auto"
filterwarnings = ["ignore:.*swigvarlink.*__module__ attribute:DeprecationWarning", "ignore:.*SwigPyObject.*__module__ attribute:DeprecationWarning", "ignore:.*SwigPyPacked.*__module__ attribute:DeprecationWarning"]
